+++ pg_metadata/install/sql/pgmetadata/20_TABLE_SEQUENCE_DEFAULT.sql
@@ -3,8 +3,8 @@ BEGIN;
 -- PostgreSQL database dump
 --
 
--- Dumped from database version 10.15 (Debian 10.15-1.pgdg100+1)
--- Dumped by pg_dump version 10.15 (Debian 10.15-1.pgdg100+1)
+-- Dumped from database version 11.7 (Debian 11.7-2.pgdg100+1)
+-- Dumped by pg_dump version 11.7 (Debian 11.7-2.pgdg100+1)
 
 SET statement_timeout = 0;
 SET lock_timeout = 0;
@@ -27,7 +27,8 @@ CREATE TABLE pgmetadata.contact (
     name text NOT NULL,
     organisation_name text NOT NULL,
     organisation_unit text,
-    email text
+    email text,
+    phone text
 );