+++ pg_metadata/install/sql/upgrade/upgrade_to_1.2.0.sql
@@ -1,149 +1,37 @@
 BEGIN;
 
--- GLOSSARY
+-- CONTACT PHONE NUMBER
 
--- additional publication frequencies
+ALTER TABLE pgmetadata.contact ADD COLUMN IF NOT EXISTS phone text;
+COMMENT ON COLUMN pgmetadata.contact.phone IS 'Phone number';
 
-INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (137, 'dataset.publication_frequency', 'QUA', 'Quarterly', 'Update data every three months', 4, NULL, NULL, NULL, NULL, NULL, NULL, 'vierteljährlich', 'Daten werden vierteljährlich aktualisiert');
-INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (138, 'dataset.publication_frequency', 'FTN', 'Fortnightly', 'Update data every two weeks', 6, NULL, NULL, NULL, NULL, NULL, NULL, 'zweiwöchentlich', 'Daten werden vierzehntägig aktualisiert');
-INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (139, 'dataset.publication_frequency', 'CON', 'Continual', 'Data is repeatedly and frequently updated', 9, NULL, NULL, NULL, NULL, NULL, NULL, 'kontinuierlich', 'Daten werden ständig aktualisiert');
-INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (140, 'dataset.publication_frequency', 'UNK', 'Unknown', 'Frequency of maintenance for the data is not known', 12, NULL, NULL, NULL, NULL, NULL, NULL, 'unbekannt', 'ein Aktualisierungsintervall ist nicht bekannt');
-
-SELECT pg_catalog.setval('pgmetadata.glossary_id_seq', 140, true);
-
-
--- new item_order for existing publication frequencies
-
-CREATE TABLE pgmetadata.t_glossary (field text, code text, item_order smallint);
-INSERT INTO pgmetadata.t_glossary (field, code, item_order)
-VALUES
-('dataset.publication_frequency', 'MON', 5),
-('dataset.publication_frequency', 'WEE', 7),
-('dataset.publication_frequency', 'DAY', 8),
-('dataset.publication_frequency', 'IRR', 10),
-('dataset.publication_frequency', 'NOP', 11)
-ON CONFLICT DO NOTHING;
-
-UPDATE pgmetadata.glossary AS g
-SET item_order = t.item_order
-FROM pgmetadata.t_glossary AS t
-WHERE g.field = t.field AND g.code = t.code;
-
-DROP TABLE pgmetadata.t_glossary;
-
-
--- DATASET
-
--- new field license_attribution
-
-ALTER TABLE pgmetadata.dataset ADD COLUMN IF NOT EXISTS license_attribution text;
-
-CREATE OR REPLACE VIEW pgmetadata.v_dataset AS
+CREATE OR REPLACE VIEW pgmetadata.v_contact AS
  WITH glossary AS (
          SELECT COALESCE(current_setting('pgmetadata.locale'::text, true), 'en'::text) AS locale,
             v_glossary.dict
            FROM pgmetadata.v_glossary
-        ), s AS (
-         SELECT d.id,
-            d.uid,
-            d.table_name,
-            d.schema_name,
-            d.title,
-            d.abstract,
-            d.categories,
-            d.themes,
-            d.keywords,
-            d.spatial_level,
-            d.minimum_optimal_scale,
-            d.maximum_optimal_scale,
-            d.publication_date,
-            d.publication_frequency,
-            d.license,
-            d.license_attribution,
-            d.confidentiality,
-            d.feature_count,
-            d.geometry_type,
-            d.projection_name,
-            d.projection_authid,
-            d.spatial_extent,
-            d.creation_date,
-            d.update_date,
-            d.data_last_update,
-            d.geom,
-            cat.cat,
-            theme.theme
-           FROM ((pgmetadata.dataset d
-             LEFT JOIN LATERAL unnest(d.categories) cat(cat) ON (true))
-             LEFT JOIN LATERAL unnest(d.themes) theme(theme) ON (true))
-          WHERE true
-          ORDER BY d.id
-        ), ss AS (
-         SELECT s.id,
-            s.uid,
-            s.table_name,
-            s.schema_name,
-            s.title,
-            s.abstract,
-            ((((glossary.dict -> 'dataset.categories'::text) -> s.cat) -> 'label'::text) ->> glossary.locale) AS cat,
-            gtheme.label AS theme,
-            s.keywords,
-            s.spatial_level,
-            ('1/'::text || s.minimum_optimal_scale) AS minimum_optimal_scale,
-            ('1/'::text || s.maximum_optimal_scale) AS maximum_optimal_scale,
-            s.publication_date,
-            ((((glossary.dict -> 'dataset.publication_frequency'::text) -> s.publication_frequency) -> 'label'::text) ->> glossary.locale) AS publication_frequency,
-            ((((glossary.dict -> 'dataset.license'::text) -> s.license) -> 'label'::text) ->> glossary.locale) AS license,
-            s.license_attribution,
-            ((((glossary.dict -> 'dataset.confidentiality'::text) -> s.confidentiality) -> 'label'::text) ->> glossary.locale) AS confidentiality,
-            s.feature_count,
-            s.geometry_type,
-            (regexp_split_to_array((rs.srtext)::text, '"'::text))[2] AS projection_name,
-            s.projection_authid,
-            s.spatial_extent,
-            s.creation_date,
-            s.update_date,
-            s.data_last_update
-           FROM glossary,
-            ((s
-             LEFT JOIN pgmetadata.theme gtheme ON ((gtheme.code = s.theme)))
-             LEFT JOIN public.spatial_ref_sys rs ON ((concat(rs.auth_name, ':', rs.auth_srid) = s.projection_authid)))
         )
- SELECT ss.id,
-    ss.uid,
-    ss.table_name,
-    ss.schema_name,
-    ss.title,
-    ss.abstract,
-    string_agg(DISTINCT ss.cat, ', '::text ORDER BY ss.cat) AS categories,
-    string_agg(DISTINCT ss.theme, ', '::text ORDER BY ss.theme) AS themes,
-    ss.keywords,
-    ss.spatial_level,
-    ss.minimum_optimal_scale,
-    ss.maximum_optimal_scale,
-    ss.publication_date,
-    ss.publication_frequency,
-    ss.license,
-    ss.confidentiality,
-    ss.feature_count,
-    ss.geometry_type,
-    ss.projection_name,
-    ss.projection_authid,
-    ss.spatial_extent,
-    ss.creation_date,
-    ss.update_date,
-    ss.data_last_update,
-    ss.license_attribution
-   FROM ss
-  GROUP BY ss.id, ss.uid, ss.table_name, ss.schema_name, ss.title, ss.abstract, ss.keywords, ss.spatial_level, ss.minimum_optimal_scale, ss.maximum_optimal_scale, ss.publication_date, ss.publication_frequency, ss.license, ss.license_attribution, ss.confidentiality, ss.feature_count, ss.geometry_type, ss.projection_name, ss.projection_authid, ss.spatial_extent, ss.creation_date, ss.update_date, ss.data_last_update;
+ SELECT d.table_name,
+    d.schema_name,
+    c.name,
+    c.organisation_name,
+    c.organisation_unit,
+    ((((glossary.dict -> 'contact.contact_role'::text) -> dc.contact_role) -> 'label'::text) ->> glossary.locale) AS contact_role,
+    dc.contact_role AS contact_role_code,
+    c.email,
+    c.phone
+   FROM glossary,
+    ((pgmetadata.dataset_contact dc
+     JOIN pgmetadata.dataset d ON ((d.id = dc.fk_id_dataset)))
+     JOIN pgmetadata.contact c ON ((dc.fk_id_contact = c.id)))
+  WHERE true
+  ORDER BY dc.id;
 
-
-
--- ADD RASTER SUPPORT
 
 DROP FUNCTION pgmetadata.calculate_fields_from_data() CASCADE;
 CREATE FUNCTION pgmetadata.calculate_fields_from_data() RETURNS trigger
     LANGUAGE plpgsql
-    AS $$
+    AS $_$
 DECLARE
     test_target_table regclass;
     target_table text;
@@ -169,12 +57,18 @@ BEGIN
         RETURN NEW;
     END IF;
 
--- Get table feature count
+    -- Date fields
+    NEW.update_date = now();
+    IF TG_OP = 'INSERT' THEN
+        NEW.creation_date = now();
+    END IF;
+
+    -- Get table feature count
     EXECUTE 'SELECT COUNT(*) FROM ' || target_table
     INTO NEW.feature_count;
     -- RAISE NOTICE 'pgmetadata - % feature_count: %', target_table, NEW.feature_count;
 
--- Check geometry properties: get data from geometry_columns and raster_columns
+    -- Check geometry properties: get data from geometry_columns and raster_columns
     EXECUTE
     ' SELECT *' ||
     ' FROM geometry_columns' ||
@@ -183,15 +77,19 @@ BEGIN
     ' LIMIT 1'
     INTO test_geom_column;
 
-    EXECUTE
-    ' SELECT *' ||
-    ' FROM raster_columns' ||
-    ' WHERE r_table_schema=' || quote_literal(NEW.schema_name) ||
-    ' AND r_table_name=' || quote_literal(NEW.table_name) ||
-    ' LIMIT 1'
-    INTO test_rast_column;
+    IF to_regclass('raster_columns') is not null THEN
+        EXECUTE
+        ' SELECT *' ||
+        ' FROM raster_columns' ||
+        ' WHERE r_table_schema=' || quote_literal(NEW.schema_name) ||
+        ' AND r_table_name=' || quote_literal(NEW.table_name) ||
+        ' LIMIT 1'
+        INTO test_rast_column;
+    ELSE
+        select null into test_rast_column;
+    END IF;
 
--- If the table has a geometry column, calculate field values
+    -- If the table has a geometry column, calculate field values
     IF test_geom_column IS NOT NULL THEN
 
         -- column name
@@ -238,28 +136,22 @@ BEGIN
         NEW.geometry_type = test_geom_column.type;
 
     ELSIF test_rast_column is not null THEN
-    
+
         -- column name
         rast_column_name = test_rast_column.r_raster_column;
         RAISE NOTICE 'pgmetadata - table % has a raster column: %', target_table, rast_column_name;
 
-        -- TODO: is this the best way to get the raster extent? Could we use ST_xmin etc. on test_rast_column.extent?
-        EXECUTE '
-            SELECT CONCAT(
-                min(ST_xmin(extent))::text, '', '',
-                max(ST_xmax(extent))::text, '', '',
-                min(ST_ymin(extent))::text, '', '',
-                max(ST_ymax(extent))::text)
-            FROM raster_columns
-            WHERE r_table_schema=' || quote_literal(NEW.schema_name) ||
-              'AND r_table_name=' || quote_literal(NEW.table_name)
-        INTO NEW.spatial_extent;
-        
-        -- convexhull from target table
-        EXECUTE '
-            SELECT ST_Transform(ST_ConvexHull("' || rast_column_name || '"), 4326)
-            FROM ' || target_table
-        INTO geom_envelop;
+        -- spatial_extent
+        EXECUTE 'SELECT CONCAT(ST_xmin($1)::text, '', '', ST_xmax($1)::text, '', '',
+                               ST_ymin($1)::text, '', '', ST_ymax($1)::text)'
+        INTO NEW.spatial_extent
+        USING test_rast_column.extent;
+
+        -- use extent (of whole table) from raster_columns catalog as envelope
+        -- (union of convexhull of all rasters (tiles) in target table is too slow for big tables)
+        EXECUTE 'SELECT ST_Transform($1, 4326)'
+        INTO geom_envelop
+        USING test_rast_column.extent;
 
         -- Test if it's not a point or a line
         IF GeometryType(geom_envelop) != 'POLYGON' THEN
@@ -270,17 +162,14 @@ BEGIN
             NEW.GEOM = geom_envelop;
         END IF;
 
-        -- projection_authid
-        EXECUTE '
-            SELECT CONCAT(s.auth_name, '':'', ST_SRID(m."' || rast_column_name || '")::text)
-            FROM ' || target_table || ' m, spatial_ref_sys s
-            WHERE s.auth_srid = ST_SRID(m."' || rast_column_name || '")
-            LIMIT 1'
-        INTO NEW.projection_authid;
+        -- projection_authid (use test_rast_column because querying table similar to vector layer is very slow)
+        EXECUTE 'SELECT CONCAT(auth_name, '':'', $1) FROM spatial_ref_sys WHERE auth_srid = $1'
+        INTO NEW.projection_authid
+        USING test_rast_column.srid;
 
         -- geometry_type
         NEW.geometry_type = 'RASTER';
-    
+
     ELSE
     -- No geometry column found: we need to erase values
             NEW.geom = NULL;
@@ -291,13 +180,243 @@ BEGIN
 
     RETURN NEW;
 END;
-$$;
+$_$;
 
 -- FUNCTION calculate_fields_from_data()
 COMMENT ON FUNCTION pgmetadata.calculate_fields_from_data() IS 'Update some fields content when updating or inserting a line in pgmetadata.dataset table.';
 
 -- restore trigger
 CREATE TRIGGER trg_calculate_fields_from_data BEFORE INSERT OR UPDATE ON pgmetadata.dataset FOR EACH ROW EXECUTE PROCEDURE pgmetadata.calculate_fields_from_data();
+
+
+
+DROP FUNCTION pgmetadata.update_postgresql_table_comment(text, text, text, text);
+CREATE FUNCTION pgmetadata.update_postgresql_table_comment(table_schema text, table_name text, table_comment text, table_type text) RETURNS boolean
+    LANGUAGE plpgsql
+    AS $$
+DECLARE
+    sql_text text;
+BEGIN
+
+    BEGIN
+        sql_text = 'COMMENT ON ' || replace(quote_literal(table_type), '''', '') || ' ' || quote_ident(table_schema) || '.' || quote_ident(table_name) || ' IS ' || quote_literal(table_comment) ;
+        EXECUTE sql_text;
+        RAISE NOTICE 'Comment updated for %', quote_ident(table_schema) || '.' || quote_ident(table_name) ;
+        RETURN True;
+    EXCEPTION WHEN OTHERS THEN
+        RAISE NOTICE 'ERROR - Failed updated comment for table %', quote_ident(table_schema) || '.' || quote_ident(table_name);
+        RETURN False;
+    END;
+
+    RETURN True;
+END;
+$$;
+
+-- FUNCTION update_postgresql_table_comment(table_schema text, table_name text, table_comment text, table_type text)
+COMMENT ON FUNCTION pgmetadata.update_postgresql_table_comment(table_schema text, table_name text, table_comment text, table_type text) IS 'Update the PostgreSQL comment of a table by giving table schema, name and comment
+Example: if you need to update the comments for all the items listed by pgmetadata.v_table_comment_from_metadata:
+
+    SELECT
+    v.table_schema,
+    v.table_name,
+    pgmetadata.update_postgresql_table_comment(
+        v.table_schema,
+        v.table_name,
+        v.table_comment,
+        v.table_type
+    ) AS comment_updated
+    FROM pgmetadata.v_table_comment_from_metadata AS v
+
+    ';
+
+
+-- Issue #115 (fix backslashes in regexp substition for HTML templates)
+
+-- generate_html_from_json(json, text)
+DROP FUNCTION IF EXISTS pgmetadata.generate_html_from_json(json, text);
+CREATE OR REPLACE FUNCTION pgmetadata.generate_html_from_json(_json_data json, _template_section text) RETURNS text
+    LANGUAGE plpgsql
+    AS $$
+DECLARE
+    item record;
+    html text;
+BEGIN
+
+    -- Get HTML template from html_template table
+    SELECT content
+    FROM pgmetadata.html_template AS h
+    WHERE True
+    AND section = _template_section
+    INTO html
+    ;
+    IF html IS NULL THEN
+        RETURN NULL;
+    END IF;
+
+    -- Get dataset item
+    -- We transpose dataset record into rows such as
+    -- col    | val
+    -- id     | 1
+    -- uid    | dfd3b73c-3cd3-40b7-b92d-aa0f625c86fe
+    -- ...
+    -- title  | My title
+    -- For each row, we search and replace the [% "col" %] by val
+    FOR item IN
+        SELECT (line.d).key AS col, Coalesce((line.d).value, '') AS val
+        FROM (
+            SELECT json_each_text(_json_data) d
+        ) AS line
+    LOOP
+        -- replace QGIS style field [% "my_field" %] by field value
+        html = regexp_replace(
+            html,
+            concat('\[% *"?', item.col, '"? *%\]'),
+            replace(item.val, '\', '\\'), -- escape backslashes in substitution string (\1...\9 refer to subexpressions)
+            'g'
+        )
+        ;
+
+    END LOOP;
+
+    RETURN html;
+
+END;
+$$;
+
+-- FUNCTION generate_html_from_json(_json_data json, _template_section text)
+COMMENT ON FUNCTION pgmetadata.generate_html_from_json(_json_data json, _template_section text) IS 'Generate HTML content for the given JSON representation of a record and a given section, based on the template stored in the pgmetadata.html_template table. Template section controlled values are "main", "contact" and "link". If the corresponding line is not found in the pgmetadata.html_template table, NULL is returned.';
+
+
+-- get_dataset_item_html_content(text, text, text)
+DROP FUNCTION IF EXISTS get_dataset_item_html_content(text, text, text);
+CREATE OR REPLACE FUNCTION pgmetadata.get_dataset_item_html_content(_table_schema text, _table_name text, _locale text) RETURNS text
+    LANGUAGE plpgsql
+    AS $$
+DECLARE
+    locale_exists boolean;
+    item record;
+    dataset_rec record;
+    sql_text text;
+    json_data json;
+    html text;
+    html_contact text;
+    html_link text;
+    html_main text;
+BEGIN
+    -- Check if dataset exists
+    SELECT *
+    FROM pgmetadata.dataset
+    WHERE True
+    AND schema_name = _table_schema
+    AND table_name = _table_name
+    LIMIT 1
+    INTO dataset_rec
+    ;
+
+    IF dataset_rec.id IS NULL THEN
+        RETURN NULL;
+    END IF;
+
+    -- Check if the _locale parameter corresponds to the available locales
+    _locale = lower(_locale);
+    SELECT _locale IN (SELECT locale FROM pgmetadata.v_locales)
+    INTO locale_exists
+    ;
+    IF NOT locale_exists THEN
+        _locale = 'en';
+    END IF;
+
+    -- Set locale
+    -- We must use EXECUTE in order to have _locale to be correctly interpreted
+    sql_text = concat('SET SESSION "pgmetadata.locale" = ', quote_literal(_locale));
+    EXECUTE sql_text;
+
+    -- Contacts
+    html_contact = '';
+    FOR json_data IN
+        WITH a AS (
+            SELECT *
+            FROM pgmetadata.v_contact
+            WHERE True
+            AND schema_name = _table_schema
+            AND table_name = _table_name
+        )
+        SELECT row_to_json(a.*)
+        FROM a
+    LOOP
+        html_contact = concat(
+            html_contact, '
+            ',
+            pgmetadata.generate_html_from_json(json_data, 'contact')
+        );
+    END LOOP;
+    -- RAISE NOTICE 'html_contact: %', html_contact;
+
+    -- Links
+    html_link = '';
+    FOR json_data IN
+        WITH a AS (
+            SELECT *
+            FROM pgmetadata.v_link
+            WHERE True
+            AND schema_name = _table_schema
+            AND table_name = _table_name
+        )
+        SELECT row_to_json(a.*)
+        FROM a
+    LOOP
+        html_link = concat(
+            html_link, '
+            ',
+            pgmetadata.generate_html_from_json(json_data, 'link')
+        );
+    END LOOP;
+    --RAISE NOTICE 'html_link: %', html_link;
+
+    -- Main
+    html_main = '';
+    WITH a AS (
+        SELECT *
+        FROM pgmetadata.v_dataset
+        WHERE True
+        AND schema_name = _table_schema
+        AND table_name = _table_name
+    )
+    SELECT row_to_json(a.*)
+    FROM a
+    INTO json_data
+    ;
+    html_main = pgmetadata.generate_html_from_json(json_data, 'main');
+    -- RAISE NOTICE 'html_main: %', html_main;
+
+    IF html_main IS NULL THEN
+        RETURN NULL;
+    END IF;
+
+    html = html_main;
+
+    -- add contacts: [% "meta_contacts" %]
+    html = regexp_replace(
+        html,
+        concat('\[% *"?meta_contacts"? *%\]'),
+        coalesce(replace(html_contact, '\', '\\'), ''), -- escape backslashes in substitution string (\1...\9 refer to subexpressions)
+        'g'
+    );
+
+    -- add links [% "meta_links" %]
+    html = regexp_replace(
+        html,
+        concat('\[% *"?meta_links"? *%\]'),
+        coalesce(replace(html_link, '\', '\\'), ''), -- escape backslashes in substitution string (\1...\9 refer to subexpressions)
+        'g'
+    );
+
+    RETURN html;
+
+END;
+$$;
+
+COMMENT ON FUNCTION pgmetadata.get_dataset_item_html_content(_table_schema text, _table_name text, _locale text) IS 'Generate the metadata HTML content for the given table and given language or NULL if no templates are stored in the pgmetadata.html_template table.';
 
 
 COMMIT;