+++ pg_metadata/pg_metadata.py
@@ -9,6 +9,10 @@ from qgis.PyQt.QtGui import QDesktopServices, QIcon
 from qgis.PyQt.QtWidgets import QAction
 from qgis.utils import iface
 
+from pg_metadata.connection_manager import (
+    store_connections,
+    validate_connections_names,
+)
 from pg_metadata.dock import PgMetadataDock
 from pg_metadata.locator import LocatorFilter
 from pg_metadata.processing.provider import PgMetadataProvider
@@ -40,15 +44,19 @@ class PgMetadata:
             # noinspection PyCallByClass,PyArgumentList
             QCoreApplication.installTranslator(self.translator)
 
+    # noinspection PyPep8Naming
     def initProcessing(self):
         """ Add the QGIS Processing provider. """
         if not self.provider:
             self.provider = PgMetadataProvider()
             QgsApplication.processingRegistry().addProvider(self.provider)
 
+    # noinspection PyPep8Naming
     def initGui(self):
         """ Build the plugin GUI. """
         self.initProcessing()
+
+        self.check_invalid_connection_names()
 
         icon = QIcon(resources_path('icons', 'icon.png'))
 
@@ -79,6 +87,34 @@ class PgMetadata:
 
         self.data_item_provider = DataItemProvider(self)
         QgsApplication.instance().dataItemProviderRegistry().addProvider(self.data_item_provider)
+
+    @staticmethod
+    def check_invalid_connection_names():
+        """ Check for invalid connection names in the QgsSettings. """
+        valid, invalid = validate_connections_names()
+        n_invalid = len(invalid)
+
+        if n_invalid == 0:
+            return
+
+        invalid_text = ', '.join(invalid)
+        msg = QMessageBox()
+        msg.setIcon(QMessageBox.Warning)
+        msg.setWindowTitle(tr('PgMetadata: Database connection(s) not available'))
+        msg.setText(tr(
+            f'{n_invalid} connection(s) listed in PgMetadata’s settings are invalid or '
+            f'no longer available: {invalid_text}'))
+        msg.setInformativeText(tr(
+            'Do you want to remove these connection(s) from the PgMetadata settings? '
+            '(You can also do this later with the “Set Connections” tool.)'))
+        msg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+        clicked = msg.exec()
+
+        if clicked == QMessageBox.Yes:
+            iface.messageBar().pushSuccess('PgMetadata', tr(f'{n_invalid} invalid connection(s) removed.'))
+            store_connections(valid)
+        if clicked == QMessageBox.No:
+            iface.messageBar().pushInfo('PgMetadata', tr(f'Keeping {n_invalid} invalid connections.'))
 
     @staticmethod
     def open_help():