+++ pg_metadata/test/base_database.py
@@ -26,7 +26,6 @@ from pg_metadata.test.base import BaseTestProcessing
 __copyright__ = "Copyright 2020, 3Liz"
 __license__ = "GPL version 3"
 __email__ = "info@3liz.org"
-__revision__ = "$Format:%H$"
 
 SCHEMA = "pgmetadata"
 
@@ -58,7 +57,7 @@ class DatabaseTestCase(BaseTestProcessing):
             "{}:create_database_structure".format(self.provider.id()), params, feedback=None,
         )
 
-        # Insert a layer with geometry
+        # Insert a layer with vector geometry
         layer = QgsVectorLayer(plugin_test_data_path('lines.geojson'), 'lines', 'ogr')
 
         uri = QgsDataSourceUri(self.connection.uri())
@@ -74,6 +73,11 @@ class DatabaseTestCase(BaseTestProcessing):
             False)
         if result[0] != 0:
             raise Exception('Layer exported did not work')
+
+        # Insert a layer with raster geometry
+        with open(plugin_test_data_path('raster.sql')) as f:
+            for line in f:
+                self.connection.executeSql(line)
 
         # Insert a layer without geometry
         layer = QgsVectorLayer(
@@ -96,7 +100,7 @@ class DatabaseTestCase(BaseTestProcessing):
         if result[0] != 0:
             raise Exception('Layer exported did not work')
 
-        for table_name in ['tabular', 'lines']:
+        for table_name in ['tabular', 'lines', 'raster']:
             # When QGIS >= 3.12, use table()
             table = [t for t in self.connection.tables(SCHEMA) if t.tableName() == table_name]
             self.assertEqual(len(table), 1)