+++ pg_metadata/test/test_database_structure.py
@@ -23,7 +23,6 @@ from pg_metadata.qgis_plugin_tools.tools.logger_processing import (
 __copyright__ = "Copyright 2020, 3Liz"
 __license__ = "GPL version 3"
 __email__ = "info@3liz.org"
-__revision__ = "$Format:%H$"
 
 VERSION = '0.0.1'
 SCHEMA = "pgmetadata"
@@ -80,10 +79,6 @@ class TestProcessing(unittest.TestCase):
         alg = "{}:upgrade_database_structure".format(provider.id())
         with self.assertRaises(QgsProcessingException):
             processing.run(alg, params, feedback=self.feedback)
-        self.assertIn(
-            "The table pgmetadata.qgis_plugin does not exist. You must first create the database structure.",
-            self.feedback.history
-        )
 
     def test_install_database(self):
         """ Test we can install the database. """