+++ pg_metadata/test/test_sql.py
@@ -96,15 +96,24 @@ class TestSql(DatabaseTestCase):
             'data_last_update': "'2020-12-25T20:35:59'::timestamp",
         }
         return_value = self._insert(dataset_feature, 'dataset', 'id')
-        link_feature = {
-            'name': "'test link'",
+        web_link_feature = {
+            'name': "'test web link'",
             'type': "'file'",
             'url': "'https://metadata.is.good'",
             'description': "''",
             'size': "0.5",
             'fk_id_dataset': "{}".format(return_value[0][0]),
         }
-        self._insert(link_feature, 'link')
+        self._insert(web_link_feature, 'link')
+        file_link_feature = {
+            'name': "'test file link'",
+            'type': "'file'",
+            'url': r"'file:///C:\Users\test\1file.txt'",
+            'description': "''",
+            'size': "0.5",
+            'fk_id_dataset': "{}".format(return_value[0][0]),
+        }
+        self._insert(file_link_feature, 'link')
 
         # Remove previous template to have a smaller one
         sql = "DELETE FROM pgmetadata.html_template WHERE section IN ('main', 'link');"
@@ -124,7 +133,7 @@ class TestSql(DatabaseTestCase):
         self._insert(html_feature, 'html_template')
         html_feature = {
             'section': "'link'",
-            'content': "'<p>[% \"name\" %] [% \"description\" %]</p><p>[% \"size\" %]</p>'",
+            'content': "'<p>[% \"name\" %] [% \"description\" %]</p><p>[% url %] [% \"size\" %]</p>'",
         }
         self._insert(html_feature, 'html_template')
 
@@ -133,7 +142,9 @@ class TestSql(DatabaseTestCase):
         )
         expected = (
             '<p>Test title</p><b>Test abstract.</b><p>2020-12-25T20:35:59</p><p>\n'
-            '            <p>test link </p><p>1</p></p><p>New test theme, test theme</p>'
+            '            <p>test web link </p><p>https://metadata.is.good 1</p>\n'
+            r'            <p>test file link </p><p>file:///C:\Users\test\1file.txt 1</p>'
+            '</p><p>New test theme, test theme</p>'
         )
         self.assertEqual(expected, result[0][0])
 
@@ -163,8 +174,8 @@ class TestSql(DatabaseTestCase):
         }
         return_value = self._insert(dataset_feature, 'dataset', 'id, uid')
 
-        link_feature = {
-            'name': "'test link'",
+        web_link_feature = {
+            'name': "'test web link'",
             'type': "'file'",
             'mime': "'pdf'",
             'url': "'https://metadata.is.good'",
@@ -172,8 +183,17 @@ class TestSql(DatabaseTestCase):
             'size': "590",
             'fk_id_dataset': "{}".format(return_value[0][0]),
         }
-        self._insert(link_feature, 'link')
-
+        self._insert(web_link_feature, 'link')
+        file_link_feature = {
+            'name': "'test file link'",
+            'type': "'file'",
+            'mime': "'plain'",
+            'url': r"'file:///C:\Users\test\1file.txt'",
+            'description': "'File description'",
+            'size': "590",
+            'fk_id_dataset': "{}".format(return_value[0][0]),
+        }
+        self._insert(file_link_feature, 'link')
         contact_feature = {
             'name': "'Jane Doe'",
             'organisation_name': "'Acme'",
@@ -254,10 +274,17 @@ class TestSql(DatabaseTestCase):
             '<foaf:mbox>bob.bob@corp.spa</foaf:mbox>'
             '</foaf:Organization></dct:publisher>'
 
-            '<dcat:distribution><dcat:Distribution><dct:title>test link</dct:title>'
+            '<dcat:distribution><dcat:Distribution><dct:title>test web link</dct:title>'
             '<dct:description>Link description</dct:description>'
             '<dcat:downloadURL>https://metadata.is.good</dcat:downloadURL>'
             '<dcat:mediaType>application/pdf</dcat:mediaType><dct:format>a file</dct:format>'
+            '<dct:bytesize>590</dct:bytesize>'
+            '<dct:license>Licence Ouverte Version 2.1</dct:license>'
+            '</dcat:Distribution></dcat:distribution>'
+            '<dcat:distribution><dcat:Distribution><dct:title>test file link</dct:title>'
+            '<dct:description>File description</dct:description>'
+            r'<dcat:downloadURL>file:///C:\Users\test\1file.txt</dcat:downloadURL>'
+            '<dcat:mediaType>text/plain</dcat:mediaType><dct:format>a file</dct:format>'
             '<dct:bytesize>590</dct:bytesize>'
             '<dct:license>Licence Ouverte Version 2.1</dct:license>'
             '</dcat:Distribution></dcat:distribution>'
@@ -327,8 +354,11 @@ class TestSql(DatabaseTestCase):
 
         # Test insert
         sql = "SELECT geometry_type, projection_authid, spatial_extent FROM pgmetadata.dataset"
-        result = self._sql(sql)
-        self.assertEqual(['LINESTRING', 'EPSG:4326', '3.854, 3.897, 43.5786, 43.622'], result[0])
+        result = self._sql(sql)[0]
+        self.assertEqual('LINESTRING', result[0])
+        self.assertEqual('EPSG:4326', result[1])
+        coordinates = [f.strip()[0:6] for f in result[2].split(',')]
+        self.assertListEqual(['3.8540', '3.8969', '43.578', '43.621'], coordinates)
 
         # Test date, creation_date is equal to update_date
         sql = "SELECT creation_date, update_date FROM pgmetadata.dataset"
@@ -339,11 +369,12 @@ class TestSql(DatabaseTestCase):
         sql = "UPDATE pgmetadata.dataset SET title = 'test lines title' WHERE table_name = 'lines'"
         self._sql(sql)
         sql = "SELECT title, geometry_type, projection_authid, spatial_extent FROM pgmetadata.dataset"
-        result = self._sql(sql)
-        self.assertEqual(
-            ['test lines title', 'LINESTRING', 'EPSG:4326', '3.854, 3.897, 43.5786, 43.622'],
-            result[0]
-        )
+        result = self._sql(sql)[0]
+        self.assertEqual('test lines title', result[0])
+        self.assertEqual('LINESTRING', result[1])
+        self.assertEqual('EPSG:4326', result[2])
+        coordinates = [f.strip()[0:6] for f in result[3].split(',')]
+        self.assertListEqual(['3.8540', '3.8969', '43.578', '43.621'], coordinates)
 
         # Test date, creation_date is not equal to update_date
         sql = "SELECT creation_date, update_date FROM pgmetadata.dataset"
@@ -378,7 +409,7 @@ class TestSql(DatabaseTestCase):
         self.assertEqual(["Test title SQL view - Test abstract SQL view. ()"], result[0])
 
     def test_trigger_calculate_fields_when_removing_geom(self):
-        """ Test the trigger on an existing table where we remove the geometry. """
+        """ Test the trigger on an existing vector table where we remove the geometry. """
 
         # Adding data
         dataset_feature = {
@@ -441,3 +472,61 @@ class TestSql(DatabaseTestCase):
         )
         result = self._sql(sql)
         self.assertListEqual(['POINT', 'EPSG:4326', '0, 0, 0, 0'], result[0])
+
+    def test_trigger_calculate_fields_raster(self):
+        """ Test if fields are correctly calculated on a raster layer (having a rast column). """
+
+        # Adding data
+        dataset_feature = {
+            'table_name': "'raster'",
+            'schema_name': "'pgmetadata'",
+            'title': "'Test title'",
+            'abstract': "'Test abstract.'",
+        }
+        self._insert(dataset_feature, 'dataset')
+
+        # Test insert
+        sql = "SELECT geometry_type, projection_authid, spatial_extent FROM pgmetadata.dataset"
+        result = self._sql(sql)
+        self.assertEqual(['RASTER', 'EPSG:25833', '300000, 300800, 5700000, 5700800'], result[0])
+
+        # Test date, creation_date is equal to update_date
+        sql = "SELECT creation_date, update_date FROM pgmetadata.dataset"
+        result = self._sql(sql)
+        self.assertEqual(result[0][0], result[0][1])
+
+        # Test update
+        sql = "UPDATE pgmetadata.dataset SET title = 'test raster title' WHERE table_name = 'raster'"
+        self._sql(sql)
+        sql = "SELECT title, geometry_type, projection_authid, spatial_extent FROM pgmetadata.dataset"
+        result = self._sql(sql)
+        self.assertEqual(
+            ['test raster title', 'RASTER', 'EPSG:25833', '300000, 300800, 5700000, 5700800'],
+            result[0]
+        )
+
+        # Test date, creation_date is not equal to update_date
+        sql = "SELECT creation_date, update_date FROM pgmetadata.dataset"
+        result = self._sql(sql)
+        self.assertNotEqual(result[0][0], result[0][1])
+
+    def test_trigger_calculate_fields_when_removing_rast(self):
+        """ Test the trigger on an existing raster table where we remove the geometry. """
+
+        # Adding data
+        dataset_feature = {
+            'table_name': "'raster'",
+            'schema_name': "'pgmetadata'",
+            'title': "'Test title'",
+            'abstract': "'Test abstract.'",
+        }
+        self._insert(dataset_feature, 'dataset')
+
+        # Test with removing the rast column
+        sql = "ALTER TABLE pgmetadata.raster DROP COLUMN rast"
+        self._sql(sql)
+        sql = "UPDATE pgmetadata.dataset SET title = 'test after drop rast column'"
+        self._sql(sql)
+        sql = "SELECT title, geometry_type, projection_authid, spatial_extent, geom FROM pgmetadata.dataset"
+        result = self._sql(sql)
+        self.assertEqual(['test after drop rast column', NULL, NULL, NULL, NULL], result[0])