+++ pg_metadata/tools.py
@@ -34,6 +34,9 @@ def icon_for_geometry_type(geometry_type: str) -> QIcon():
     elif geometry_type == 'RASTER':
         return QgsLayerItem.iconRaster()
 
+    elif geometry_type == 'RASTER':
+        return QgsLayerItem.iconRaster()
+
     # Default icon
     return QIcon(resources_path('icons', 'icon.png'))